Modify existing example
