--- examples/react/urql-string-document-mode/codegen.ts
@@ -1,18 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { type CodegenConfig } from '@graphql-codegen/cli';
-
-const config: CodegenConfig = {
-  schema: 'https://swapi-graphql.netlify.app/.netlify/functions/index',
-  documents: ['src/**/*.tsx', '!src/gql/**/*'],
-  generates: {
-    './src/gql/': {
-      preset: 'client',
-      config: {
-        documentMode: 'string',
-      },
-    },
-  },
-  hooks: { afterAllFileWrite: ['prettier --write'] },
-};
-
-export default config;
