--- examples/react/urql-string-document-mode/cypress.config.ts
@@ -1,10 +1,0 @@
-/* eslint-disable import/no-extraneous-dependencies */
-import { defineConfig } from 'cypress';
-
-export default defineConfig({
-  e2e: {
-    setupNodeEvents(_on, _config) {
-      // implement node event listeners here
-    },
-  },
-});
