--- examples/react/urql-string-document-mode/cypress/e2e/end2end.cy.ts
@@ -1,6 +1,0 @@
-describe('template spec', () => {
-  it('renders everything correctly', () => {
-    cy.visit('http://localhost:3000');
-    cy.get('h3').should('contain', 'A New Hope');
-  });
-});
