--- examples/react/urql-string-document-mode/cypress/support/commands.ts
@@ -1,1 +1,0 @@
-/// <reference types="cypress" />
