--- examples/react/urql-string-document-mode/cypress/support/e2e.ts
@@ -1,1 +1,0 @@
-import './commands';
