--- examples/react/urql-string-document-mode/package.json
@@ -1,30 +1,0 @@
-{
-  "name": "example-react-urql-string-document-node",
-  "version": "0.1.0",
-  "private": true,
-  "dependencies": {
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "urql": "^3.0.0"
-  },
-  "devDependencies": {
-    "@types/react": "^18.0.17",
-    "@types/react-dom": "^18.0.10",
-    "@graphql-codegen/cli": "^3.2.2",
-    "@graphql-codegen/client-preset": "^2.1.1",
-    "@vitejs/plugin-react": "^3.1.0",
-    "typescript": "4.9.5",
-    "serve": "14.2.0",
-    "cypress": "12.6.0",
-    "start-server-and-test": "2.0.0",
-    "vite": "^4.1.0"
-  },
-  "scripts": {
-    "dev": "vite",
-    "build": "vite build",
-    "start": "serve -s dist",
-    "test": "cypress run",
-    "test:end2end": "start-server-and-test start http://localhost:3000 test",
-    "codegen": "graphql-codegen --config codegen.ts"
-  }
-}
