--- examples/react/urql-string-document-mode/src/Film.tsx
@@ -1,26 +1,0 @@
-import { FragmentType, useFragment } from './gql/fragment-masking';
-import { graphql } from './gql';
-
-export const FilmFragment = graphql(/* GraphQL */ `
-  fragment FilmItem on Film {
-    id
-    title
-    releaseDate
-    producers
-  }
-`);
-
-const Film = (props: {
-  /* tweet property has the correct type 🎉 */
-  film: FragmentType<typeof FilmFragment>;
-}) => {
-  const film = useFragment(FilmFragment, props.film);
-  return (
-    <div>
-      <h3>{film.title}</h3>
-      <p>{film.releaseDate}</p>
-    </div>
-  );
-};
-
-export default Film;
