--- examples/react/urql-string-document-mode/src/gql/index.ts
@@ -1,2 +1,0 @@
-export * from './fragment-masking';
-export * from './gql';
