+++ examples/react/urql/codegen.ts
@@ -7,6 +7,9 @@ const config: CodegenConfig = {
   generates: {
     './src/gql/': {
       preset: 'client',
+      config: {
+        documentMode: 'string',
+      },
     },
   },
   hooks: { afterAllFileWrite: ['prettier --write'] },
