+++ examples/react/urql/src/App.tsx
@@ -1,11 +1,12 @@
-import React from 'react';
+import { TypedDocumentString } from '@graphql-typed-document-node/core';
+import { AnyVariables, OperationContext, RequestPolicy, TypedDocumentNode, useQuery, UseQueryResponse } from 'urql';
+import type { DocumentNode } from 'graphql';
 import './App.css';
 import Film from './Film';
 import { graphql } from './gql';
-import { useQuery } from 'urql';
 
 const allFilmsWithVariablesQueryDocument = graphql(/* GraphQL */ `
-  query allFilmsWithVariablesQuery($first: Int!) {
+  query allFilmsWithVariablesQuery199($first: Int!) {
     allFilms(first: $first) {
       edges {
         node {
@@ -15,6 +16,32 @@ const allFilmsWithVariablesQueryDocument = graphql(/* GraphQL */ `
     }
   }
 `);
+
+declare module 'urql' {
+  // @ts-expect-error this is just temporary until we update types in urql
+  export type UseQueryArgs<Variables extends AnyVariables = AnyVariables, Data = any> = {
+    query: string | DocumentNode | TypedDocumentNode<Data, Variables> | TypedDocumentString<Data, Variables>;
+    requestPolicy?: RequestPolicy;
+    context?: Partial<OperationContext>;
+    pause?: boolean;
+  } & (Variables extends void
+    ? {
+        variables?: Variables;
+      }
+    : Variables extends {
+        [P in keyof Variables]: Variables[P] | null;
+      }
+    ? {
+        variables?: Variables;
+      }
+    : {
+        variables: Variables;
+      });
+
+  export function useQuery<Data = any, Variables extends AnyVariables = AnyVariables>(
+    args: UseQueryArgs<Variables, Data>
+  ): UseQueryResponse<Data, Variables>;
+}
 
 function App() {
   const [{ data }] = useQuery({
